great patch by craig@adaptivetrade.com (Craig Kelley)  -- Now rows are interperated properly as ushorts versus shorts  (java ints)
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/poi/trunk@352609 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/CellValueRecordInterface.java
@@ -81,7 +81,8 @@ public interface CellValueRecordInterface
      * @return the row
      */
 
-    public short getRow();
+    //public short getRow();
+    public int getRow();
 
     /**
      * get the column this cell defines within the row
@@ -96,7 +97,8 @@ public interface CellValueRecordInterface
      * @param row the row this cell occurs within
      */
 
-    public void setRow(short row);
+    //public void setRow(short row);
+    public void setRow(int row);
 
     /**
      * set the column this cell defines within the row
